requirements.txt
PyQt6
pillow
exifread
pyinstaller